% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_callout_question_lines}
\alias{build_callout_question_lines}
\title{Build callout-style MCQ lines}
\usage{
build_callout_question_lines(question, answers, language = "en")
}
\description{
Build callout-style MCQ lines
}
\keyword{internal}
