% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_default_setup_chunk}
\alias{build_default_setup_chunk}
\title{Build default setup chunk when none is present}
\usage{
build_default_setup_chunk(language, seed = NULL)
}
\description{
Build default setup chunk when none is present
}
\keyword{internal}
