% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_engine_chunk_header}
\alias{build_engine_chunk_header}
\title{Build generic engine chunk header}
\usage{
build_engine_chunk_header(engine, label = NULL, options = character())
}
\description{
Build generic engine chunk header
}
\keyword{internal}
