% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_exercise_lines}
\alias{build_exercise_lines}
\title{Build exercise lines from source R chunk}
\usage{
build_exercise_lines(format, exercise_label, source_chunk, language, tags)
}
\description{
Build exercise lines from source R chunk
}
\keyword{internal}
