% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_learnr_question_lines}
\alias{build_learnr_question_lines}
\title{Build learnr::question chunk lines}
\usage{
build_learnr_question_lines(question, answers, label, allow_retry = TRUE)
}
\description{
Build learnr::question chunk lines
}
\keyword{internal}
