% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_mcq_for_chunk_mode}
\alias{build_mcq_for_chunk_mode}
\title{Build MCQ lines according to selected mode}
\usage{
build_mcq_for_chunk_mode(
  source_chunk,
  format,
  mcq_label_base,
  registry,
  language,
  question_bank,
  mcq_source,
  index
)
}
\description{
Build MCQ lines according to selected mode
}
\keyword{internal}
