% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_output_yaml}
\alias{build_output_yaml}
\title{Build YAML header lines for target format}
\usage{
build_output_yaml(source_yaml, format, language = "en")
}
\description{
Build YAML header lines for target format
}
\keyword{internal}
