% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_r_chunk_header}
\alias{build_r_chunk_header}
\title{Build knitr R chunk header}
\usage{
build_r_chunk_header(label = NULL, options = character())
}
\description{
Build knitr R chunk header
}
\keyword{internal}
