% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{build_solution_lines}
\alias{build_solution_lines}
\title{Build solution lines from source R chunk}
\usage{
build_solution_lines(format, solution_label, source_chunk, language)
}
\description{
Build solution lines from source R chunk
}
\keyword{internal}
