% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tutorial.R
\name{check_tutorial}
\alias{check_tutorial}
\title{Render a generated tutorial and return render status}
\usage{
check_tutorial(file, language = c("en", "fr"))
}
\arguments{
\item{file}{Path to tutorial `.Rmd` to render.}

\item{language}{Message language (`"en"` or `"fr"`).}
}
\value{
Invisibly, the render output path or the error object.
}
\description{
Render a generated tutorial and return render status
}
\examples{
\dontrun{
check_tutorial("analysis-tutorial.Rmd")
}
}
