% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tutorial.R
\name{check_tutorial_render}
\alias{check_tutorial_render}
\title{Render check implementation for generated learnr tutorials}
\usage{
check_tutorial_render(file, language = "en")
}
\description{
Render check implementation for generated learnr tutorials
}
\keyword{internal}
