% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_folder.R
\name{convert_folder}
\alias{convert_folder}
\title{Convert all .Rmd/.qmd files in a folder}
\usage{
convert_folder(dir = ".", pattern = "\\\\.(Rmd|qmd)$", recursive = FALSE, ...)
}
\arguments{
\item{dir}{Path to the directory containing files.}

\item{pattern}{Regex pattern for files to include. Default matches .Rmd and .qmd.}

\item{recursive}{Logical, search recursively?}

\item{...}{Arguments passed to \code{convert_to_tutorial}.}
}
\value{
Invisibly, a vector of converted file paths.
}
\description{
Convert all .Rmd/.qmd files in a folder
}
