% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{derive_manifest_items}
\alias{derive_manifest_items}
\title{Derive LMS items from parsed chunks}
\usage{
derive_manifest_items(blocks, include_solutions = TRUE)
}
\description{
Derive LMS items from parsed chunks
}
\keyword{internal}
