% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{detect_unknown_tags}
\alias{detect_unknown_tags}
\title{Detect unknown tutorizeR tags from comment lines}
\usage{
detect_unknown_tags(lines)
}
\description{
Detect unknown tutorizeR tags from comment lines
}
\keyword{internal}
