% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{estimate_completion_minutes}
\alias{estimate_completion_minutes}
\title{Estimate expected completion time in minutes}
\usage{
estimate_completion_minutes(stats)
}
\description{
Estimate expected completion time in minutes
}
\keyword{internal}
