% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{find_i18n_file}
\alias{find_i18n_file}
\title{Locate an i18n YAML file for the given language}
\usage{
find_i18n_file(language)
}
\description{
Locate an i18n YAML file for the given language
}
\keyword{internal}
