% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{generate_simple_diff}
\alias{generate_simple_diff}
\title{Build a simple line diff between source and output}
\usage{
generate_simple_diff(source_lines, output_lines)
}
\description{
Build a simple line diff between source and output
}
\keyword{internal}
