% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{get_i18n_dictionary}
\alias{get_i18n_dictionary}
\title{Load i18n dictionary from YAML with cache}
\usage{
get_i18n_dictionary(language)
}
\description{
Load i18n dictionary from YAML with cache
}
\keyword{internal}
