% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{has_lint_errors}
\alias{has_lint_errors}
\title{Check whether lint report contains blocking errors}
\usage{
has_lint_errors(report)
}
\description{
Check whether lint report contains blocking errors
}
\keyword{internal}
