% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{is_mcq_block}
\alias{is_mcq_block}
\title{Determine if block is explicit MCQ definition block}
\usage{
is_mcq_block(block)
}
\description{
Determine if block is explicit MCQ definition block
}
\keyword{internal}
