% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{is_mcq_ref_block}
\alias{is_mcq_ref_block}
\title{Determine if block is MCQ reference block}
\usage{
is_mcq_ref_block(block)
}
\description{
Determine if block is MCQ reference block
}
\keyword{internal}
