% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{launch_tutorizeR_addin}
\alias{launch_tutorizeR_addin}
\title{RStudio Addin: Convert Active Document to Tutorial}
\usage{
launch_tutorizeR_addin()
}
\description{
Interactively converts the currently open document in RStudio.
}
