% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{launch_tutorizeR_folder_addin}
\alias{launch_tutorizeR_folder_addin}
\title{RStudio Addin: Convert Folder to Tutorial}
\usage{
launch_tutorizeR_folder_addin()
}
\description{
Interactively selects a folder and converts all contained .Rmd/.qmd files.
}
