% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{lookup_i18n_key}
\alias{lookup_i18n_key}
\title{Lookup nested translation key in dictionary}
\usage{
lookup_i18n_key(dict, key)
}
\description{
Lookup nested translation key in dictionary
}
\keyword{internal}
