% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{make_unique_label}
\alias{make_unique_label}
\title{Register and generate unique deterministic labels}
\usage{
make_unique_label(base, registry)
}
\description{
Register and generate unique deterministic labels
}
\keyword{internal}
