% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{new_tutorize_folder_report}
\alias{new_tutorize_folder_report}
\title{Build folder conversion report object}
\usage{
new_tutorize_folder_report(results)
}
\description{
Build folder conversion report object
}
\keyword{internal}
