% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{new_tutorize_lint_report}
\alias{new_tutorize_lint_report}
\title{Construct a lint report object}
\usage{
new_tutorize_lint_report(source, findings, context = "source")
}
\description{
Construct a lint report object
}
\keyword{internal}
