% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_bank.R
\name{normalize_question_entries}
\alias{normalize_question_entries}
\title{Normalize root object to list of question entries}
\usage{
normalize_question_entries(raw)
}
\description{
Normalize root object to list of question entries
}
\keyword{internal}
