% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tutorial.R
\name{parse_chunks}
\alias{parse_chunks}
\title{Transform source lines into converted tutorial body}
\usage{
parse_chunks(lines, assessment = "code", format = "learnr")
}
\description{
Transform source lines into converted tutorial body
}
\keyword{internal}
