% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_input_document}
\alias{parse_input_document}
\title{Parse an input R Markdown / Quarto document into structured blocks}
\usage{
parse_input_document(input_file)
}
\description{
Parse an input R Markdown / Quarto document into structured blocks
}
\keyword{internal}
