% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_markdown_blocks}
\alias{parse_markdown_blocks}
\title{Parse markdown body into text/chunk blocks}
\usage{
parse_markdown_blocks(lines)
}
\description{
Parse markdown body into text/chunk blocks
}
\keyword{internal}
