% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{parse_mcq_ref_yaml}
\alias{parse_mcq_ref_yaml}
\title{Parse YAML from a tutorizeR-mcq-ref block}
\usage{
parse_mcq_ref_yaml(block)
}
\description{
Parse YAML from a tutorizeR-mcq-ref block
}
\keyword{internal}
