% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_bank.R
\name{parse_question_bank_file}
\alias{parse_question_bank_file}
\title{Parse a single question-bank file}
\usage{
parse_question_bank_file(file)
}
\description{
Parse a single question-bank file
}
\keyword{internal}
