% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_tutorizer_tags}
\alias{parse_tutorizer_tags}
\title{Parse teacher tags declared in code comments}
\usage{
parse_tutorizer_tags(lines)
}
\description{
Parse teacher tags declared in code comments
}
\keyword{internal}
