% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{print.tutorize_lint_report}
\alias{print.tutorize_lint_report}
\title{Print method for lint reports}
\usage{
\method{print}{tutorize_lint_report}(x, ...)
}
\arguments{
\item{x}{A `tutorize_lint_report` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print method for lint reports
}
