% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{print.tutorize_lms_manifest}
\alias{print.tutorize_lms_manifest}
\title{Print method for LMS manifest exports}
\usage{
\method{print}{tutorize_lms_manifest}(x, ...)
}
\arguments{
\item{x}{A `tutorize_lms_manifest` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print method for LMS manifest exports
}
