% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{register_label}
\alias{register_label}
\title{Register external label without modification}
\usage{
register_label(label, registry)
}
\description{
Register external label without modification
}
\keyword{internal}
