% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{sanitize_label}
\alias{sanitize_label}
\title{Sanitize chunk labels to portable deterministic strings}
\usage{
sanitize_label(x)
}
\description{
Sanitize chunk labels to portable deterministic strings
}
\keyword{internal}
