% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_bank.R
\name{select_questions_from_bank}
\alias{select_questions_from_bank}
\title{Select question(s) from a loaded question bank}
\usage{
select_questions_from_bank(
  bank,
  ids = NULL,
  n = NULL,
  strategy = c("ordered", "random"),
  tags = NULL,
  difficulty = NULL,
  language = NULL
)
}
\description{
Select question(s) from a loaded question bank
}
\keyword{internal}
