% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{split_front_matter}
\alias{split_front_matter}
\title{Split optional YAML front matter from body lines}
\usage{
split_front_matter(lines)
}
\description{
Split optional YAML front matter from body lines
}
\keyword{internal}
