% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_mcq_ref_block}
\alias{transform_mcq_ref_block}
\title{Convert a parsed MCQ reference block to output lines}
\usage{
transform_mcq_ref_block(block, format, index, question_bank, language = "en")
}
\description{
Convert a parsed MCQ reference block to output lines
}
\keyword{internal}
