% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_setup_chunk}
\alias{transform_setup_chunk}
\title{Transform existing setup chunk and inject required lines for learnr}
\usage{
transform_setup_chunk(block, format, language, seed, registry)
}
\description{
Transform existing setup chunk and inject required lines for learnr
}
\keyword{internal}
