% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{trim_trailing_blank_lines}
\alias{trim_trailing_blank_lines}
\title{Trim trailing blank lines from final output}
\usage{
trim_trailing_blank_lines(lines)
}
\description{
Trim trailing blank lines from final output
}
\keyword{internal}
