% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{validate_mcq_definition}
\alias{validate_mcq_definition}
\title{Validate expected MCQ fields}
\usage{
validate_mcq_definition(definition, index)
}
\description{
Validate expected MCQ fields
}
\keyword{internal}
