% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{validate_mcq_ref_definition}
\alias{validate_mcq_ref_definition}
\title{Validate expected MCQ ref fields}
\usage{
validate_mcq_ref_definition(definition, language = "en")
}
\description{
Validate expected MCQ ref fields
}
\keyword{internal}
