% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_tutorial.R
\name{parse_chunks}
\alias{parse_chunks}
\title{Transform source chunks into learnr exercises + solutions}
\usage{
parse_chunks(lines, add_mcq = FALSE)
}
\arguments{
\item{lines}{Character vector, output of \code{readLines()}.}

\item{add_mcq}{Logical, add MCQ skeletons?}
}
\value{
Character vector representing the transformed body.
}
\description{
\strong{Rule}: any chunk whose label contains the word
\code{setup} (case-insensitive) is copied verbatim and \emph{not}
converted.
}
\keyword{internal}
